# perf: rotation axis aware quaternion multiplication 