--- a/cpp/plainmp/kinematics/algorithm.cpp
+++ b/cpp/plainmp/kinematics/algorithm.cpp
@@ -145,18 +145,17 @@ void KinematicModel<Scalar>::set_joint_angles_impl(
         tf_plink_to_hlink.quat() = tf_pjoint_to_hlink_quat;
       } else {
         tf_plink_to_hlink.quat() =
-            (joint_attach_rot_axes_[joint_id] ==
-             plainmp::spatial::RotAxis::NoRotation)
+            (joint_attach_rot_axes_[joint_id] == RotAxis::NoRotation)
                 ? tf_pjoint_to_hlink_quat
                 : tf_plink_to_pjoint_quat * tf_pjoint_to_hlink_quat;
       }
       tf_plink_to_hlink.trans() = tf_plink_to_pjoint_trans;
-      tf_plink_to_hlink.rotation_type_ = plainmp::spatial::RotAxis::General;
+      tf_plink_to_hlink.rotation_type_ = RotAxis::General;
     } else {
       Vector3&& trans = joint_axes_[joint_id] * joint_angles[i];
       tf_plink_to_hlink.trans() = tf_plink_to_pjoint_trans + trans;
       tf_plink_to_hlink.quat().setIdentity();
-      tf_plink_to_hlink.rotation_type_ = plainmp::spatial::RotAxis::NoRotation;
+      tf_plink_to_hlink.rotation_type_ = RotAxis::NoRotation;
     }
   }
 }