--- a/cpp/plainmp/kinematics/algorithm.cpp
+++ b/cpp/plainmp/kinematics/algorithm.cpp
@@ -150,12 +150,12 @@ void KinematicModel<Scalar>::set_joint_angles_impl(
                 : tf_plink_to_pjoint_quat * tf_pjoint_to_hlink_quat;
       }
       tf_plink_to_hlink.trans() = tf_plink_to_pjoint_trans;
-      tf_plink_to_hlink.rotation_type_ = RotAxis::General;
+      tf_plink_to_hlink.rot_axis_ = RotAxis::General;
     } else {
       Vector3&& trans = joint_axes_[joint_id] * joint_angles[i];
       tf_plink_to_hlink.trans() = tf_plink_to_pjoint_trans + trans;
       tf_plink_to_hlink.quat().setIdentity();
-      tf_plink_to_hlink.rotation_type_ = RotAxis::NoRotation;
+      tf_plink_to_hlink.rot_axis_ = RotAxis::NoRotation;
     }
   }
 }