--- a/cpp/plainmp/kinematics/kinematics.cpp
+++ b/cpp/plainmp/kinematics/kinematics.cpp
@@ -175,12 +175,9 @@ void KinematicModel<Scalar>::init_transform_cache(
                           std::is_same<Scalar, float>::value,
                       "Scalar must be double or float");
       }
-
-      bool is_approx_identity =
-          (std::abs(eigen_joint_orientation.w() - 1.0) < 1e-6);
-      // tf_plink_to_hlink_cache_[hid].is_quat_identity_ = is_approx_identity;
-      tf_plink_to_hlink_cache_[hid].rotation_type_ =
-          plainmp::spatial::RotationType::Identity;
+      auto rot_type =
+          plainmp::spatial::determine_rotation_type(eigen_joint_orientation);
+      tf_plink_to_hlink_cache_[hid].rotation_type_ = rot_type;
     }
   }
 }