conclusion: This complicated change is not worth adding. 

bench using fetch domain

### before
median: 0.5283355712890625 ms
median: 0.5311965942382812 ms
median: 0.5335807800292969 ms
median: 0.5271434783935547 ms
median: 0.5335807800292969 ms
median: 0.5346536636352539 ms
median: 0.5435943603515625 ms
median: 0.5328655242919922 ms
median: 0.5428791046142578 ms
median: 0.531315803527832 ms
median: 0.5431175231933594 ms
median: 0.52642822265625 ms
median: 0.5488395690917969 ms
median: 0.5502700805664062 ms
median: 0.5563497543334961 ms


### aftetr
median: 0.5593299865722656 ms
median: 0.5812644958496094 ms
median: 0.6077289581298828 ms
median: 0.6316900253295898 ms
median: 0.5543231964111328 ms
median: 0.5981922149658203 ms
median: 0.5927085876464844 ms
median: 0.5688667297363281 ms
median: 0.5407333374023438 ms
median: 0.5487203598022461 ms
median: 0.5431175231933594 ms
median: 0.5476474761962891 ms
median: 0.5451440811157227 ms
median: 0.5477666854858398 ms
median: 0.579833984375 ms